feat: resolve comment
