+++ src/core/agent/services/connectionService.types.ts
@@ -1,7 +1,35 @@
+import { IpexMessage } from "../agent.types";
+
 enum KeriaContactKeyPrefix {
   CONNECTION_NOTE = "note:",
   HISTORY_IPEX = "history:ipex:",
   HISTORY_REVOKE = "history:revoke:",
 }
 
-export { KeriaContactKeyPrefix };
+interface IpexHistoryItem {
+  id: string;
+  credentialType: string;
+  content: IpexMessage;
+  historyType: ConnectionHistoryType;
+  timestamp: Date;
+  connectionId: string;
+}
+
+interface KeriaContact {
+  alias: string;
+  id: string;
+  oobi: string;
+  challenges: string[];
+  wellKnowns: string[];
+}
+
+enum ConnectionHistoryType {
+  CREDENTIAL_ISSUANCE,
+  CREDENTIAL_REQUEST_PRESENT,
+  CREDENTIAL_REQUEST_AGREE,
+  CREDENTIAL_REVOKED,
+}
+
+export { ConnectionHistoryType, KeriaContactKeyPrefix };
+
+export type { KeriaContact, IpexHistoryItem, IpexMessage };
